Начал описывать процесс построения мира из фалов

+++ code/ComeForBrains/ComeForBrains/Core/Building/Items/InfectionKillerBuilder.cs
@@ -2,7 +2,7 @@ namespace ComeForBrains.Core.Building.Items;
 
 public class InfectionKillerBuilder : ItemBuilder
 {
-    public int SuccessChance { get; set; }
+    public double SuccessChance { get; set; }
     public double HealthDamage { get; set; }
     public double EffectiveTime { get; set; }
 }
