+++ code/ComeForBrains/ComeForBrains/Service/IRandom.cs
@@ -2,7 +2,7 @@ namespace ComeForBrains.Service;
 
 public interface IRandom
 {
-    public bool Try(int chance);
+    public bool Try(double chance);
     public double NextDouble(double max);
     public double NextDouble(double min, double max);
 }
